Issue #25: Rename ReplaceConfigToken to ProcessConfigFile.
