+++ tests/fixtures/tasks/process-config-file/extract.yml
