+++ tests/fixtures/tasks/process-config-file/task.yml
